% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pizza_rat.R
\name{pizzatime}
\alias{pizzatime}
\title{Pizzatime}
\usage{
pizzatime(day, money, weight)
}
\arguments{
\item{day}{= Day of the week. pizza is only to be eaten on tuesdays.}

\item{money}{= Boolean. A character value of your financial status, where "rich" is positive.}

\item{weight}{= Body weight in pounds. Calibrated to the authors average weight.}
}
\value{
If parameters are met ("tuesday", "rich", <=177), returns a message suggesting that it's pizza time. If any arguments fail, then better luck next time.
}
\description{
Pizzatime
}
\examples{
pizzatime("tuesday", "rich", 165)
}
