% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_function_practice.R
\name{vol_fun}
\alias{vol_fun}
\title{Title}
\usage{
vol_fun(L, W, H)
}
\arguments{
\item{L}{Length}

\item{W}{Width}

\item{H}{Height}
}
\value{
Volume
}
\description{
Title
}
\examples{
vol_fun(3, 3, 3)
 V = 3 * 3 * 3
 V = 27

}
